Massive update of Kerberos compatibility code
If krb5-config produces results that don't work for Kerberos probes,
fall back on manual library probing rather than just failing.

If KRB5_CONFIG was explicitly set in the environment, don't use a
different krb5-config based on --with-krb5.  If krb5-config isn't
executable, don't use it.  This allows one to force library probing by
setting KRB5_CONFIG to point to a nonexistent file.

Sanity-check the results of krb5-config before proceeding and error
out in configure if they don't work.

Add separate --with-krb5-lib and --with-krb5-include configure options
to set the library and include directories independently, and handle
lib64 systems better and more automatically.

Import the new Kerberos compatibility layer from rra-c-util and
supplement it with the principal manipulation functions needed here.
Take advantage of it to massively simplify the code.