Change module name to krb5_sync, add more configuration docs
The name of the plugin is now krb5_sync.so instead of passwd_update.so
and is installed under /usr/local/lib/krb5/plugins by default.  The
KDC configuration for the name of the module to load will need to
change accordingly.

Add configuration documentation for Heimdal and MIT post 1.9 to README.