Add ad_queue_only to force queuing of all changes
Add a new boolean krb5.conf option, ad_queue_only, which, if set to
true, forces all changes to be queued even if there are no conflicting
changes already queued.  The changes can then be processed later with
krb5-sync-backend.  This can be useful if real-time updates to Active
Directory cause performance issues in kadmind or kpasswdd.  kpasswd
clients in particular are often intolerant of delays.